:truck: [#467] Move Error components into own folder
